Changed Volume long to decimal
Indicators.Volume type changed
ObvResult would have needed to be changed to match, didn't want to take that step since it was oiut of the scope of the issue so I cast Volume back to long

+++ Indicators/Obv/Obv.cs
@@ -29,11 +29,11 @@ namespace Skender.Stock.Indicators
                 }
                 else if (h.Close > prevClose)
                 {
-                    obv += h.Volume;
+                    obv += (long)h.Volume;
                 }
                 else if (h.Close < prevClose)
                 {
-                    obv -= h.Volume;
+                    obv -= (long)h.Volume;
                 }
 
                 ObvResult result = new ObvResult
